+++ tests/ui/compile-fail/pin_data/tuple_struct_missing_pin_phantom.rs
@@ -1,0 +1,8 @@
+#![deny(warnings)]
+
+use pin_init::*;
+
+#[pin_data]
+struct Tuple<T>(T, core::marker::PhantomPinned);
+
+fn main() {}