+++ tests/ui/compile-fail/pin_data/tuple_struct_pinned_field_not_unpin.rs
@@ -1,0 +1,11 @@
+use core::marker::PhantomPinned;
+use pin_init::*;
+
+#[pin_data]
+struct Tuple<T>(#[pin] PhantomPinned, T);
+
+fn assert_unpin<T: Unpin>() {}
+
+fn main() {
+    assert_unpin::<Tuple<usize>>();
+}